Share one HLS manifest rewriter between the two proxies
The stream proxy and cast proxy each carried a near-identical manifest
line-walker and URI="..." tag parser — two divergent copies of the
trickiest parsing in the proxy layer. proxy_common now owns a single
rewriter parameterized by a URL mapper closure (returning None leaves a
reference unrewritten, which expresses the cast proxy's origin filter),
plus the shared is_m3u8 detection.
