+++ src-tauri/src/engine/proxy_common.rs
@@ -1,6 +1,120 @@
 //! Helpers shared by the playback stream proxy, the cast proxy, and the
 //! checker for talking to untrusted upstream media servers.
 
+use url::Url;
+
+/// Whether a response is an HLS manifest, by content type or .m3u8 path.
+pub fn is_m3u8_response(content_type: &str, url: &str) -> bool {
+    let ct = content_type.to_lowercase();
+    if ct.contains("application/vnd.apple.mpegurl") || ct.contains("application/x-mpegurl") {
+        return true;
+    }
+    if let Ok(parsed) = Url::parse(url) {
+        if parsed.path().to_lowercase().ends_with(".m3u8") {
+            return true;
+        }
+    }
+    false
+}
+
+/// Rewrite URIs in an HLS manifest so they route back through a proxy.
+///
+/// Handles bare URI lines (segment and playlist references) and URI="..."
+/// attributes in #EXT-X-MAP, #EXT-X-KEY, #EXT-X-MEDIA, #EXT-X-SESSION-KEY.
+/// `map_uri` receives each URI resolved against the manifest base and returns
+/// the replacement, or `None` to leave the reference unrewritten (e.g. an
+/// out-of-origin target the caller refuses to proxy).
+pub fn rewrite_hls_manifest(
+    body: &str,
+    base_url: &str,
+    map_uri: &dyn Fn(&Url) -> Option<String>,
+) -> String {
+    let base = match Url::parse(base_url) {
+        Ok(u) => u,
+        Err(_) => return body.to_string(),
+    };
+
+    let mut output = String::with_capacity(body.len());
+    for line in body.lines() {
+        let trimmed = line.trim();
+
+        if trimmed.is_empty() {
+            output.push('\n');
+            continue;
+        }
+
+        if trimmed.starts_with('#') {
+            let upper = trimmed.to_ascii_uppercase();
+            if (upper.starts_with("#EXT-X-MAP:")
+                || upper.starts_with("#EXT-X-KEY:")
+                || upper.starts_with("#EXT-X-MEDIA:")
+                || upper.starts_with("#EXT-X-SESSION-KEY:"))
+                && trimmed.contains("URI=")
+            {
+                output.push_str(&rewrite_tag_uri(trimmed, &base, map_uri));
+            } else {
+                output.push_str(line);
+            }
+            output.push('\n');
+            continue;
+        }
+
+        // Non-comment, non-empty line: a URI reference
+        match base.join(trimmed).ok().and_then(|u| map_uri(&u)) {
+            Some(replacement) => output.push_str(&replacement),
+            None => output.push_str(line),
+        }
+        output.push('\n');
+    }
+
+    output
+}
+
+/// Rewrite the URI="..." value inside an HLS tag line.
+fn rewrite_tag_uri(line: &str, base: &Url, map_uri: &dyn Fn(&Url) -> Option<String>) -> String {
+    let upper = line.to_ascii_uppercase();
+    let Some(uri_pos) = upper.find("URI=") else {
+        return line.to_string();
+    };
+
+    let after_uri_eq = &line[uri_pos + 4..];
+
+    let (quote, uri_start, uri_end) = if after_uri_eq.starts_with('"') {
+        let inner = &after_uri_eq[1..];
+        let end = inner.find('"').unwrap_or(inner.len());
+        (Some('"'), 1, 1 + end)
+    } else if after_uri_eq.starts_with('\'') {
+        let inner = &after_uri_eq[1..];
+        let end = inner.find('\'').unwrap_or(inner.len());
+        (Some('\''), 1, 1 + end)
+    } else {
+        let end = after_uri_eq
+            .find(|c: char| c == ',' || c.is_whitespace())
+            .unwrap_or(after_uri_eq.len());
+        (None, 0, end)
+    };
+
+    let original_uri = &after_uri_eq[uri_start..uri_end];
+    let Some(new_uri) = base.join(original_uri).ok().and_then(|u| map_uri(&u)) else {
+        return line.to_string();
+    };
+
+    let mut result = String::with_capacity(line.len() + new_uri.len());
+    result.push_str(&line[..uri_pos + 4]);
+    if let Some(q) = quote {
+        result.push(q);
+        result.push_str(&new_uri);
+        result.push(q);
+    } else {
+        result.push_str(&new_uri);
+    }
+    let remainder_offset = uri_pos + 4 + uri_end + if quote.is_some() { 1 } else { 0 };
+    if remainder_offset < line.len() {
+        result.push_str(&line[remainder_offset..]);
+    }
+    result
+}
+
 /// Read an HTTP request head from a raw socket until the `\r\n\r\n`
 /// terminator, bounded at `max_bytes`. A single `read()` is not enough: the
 /// request line and headers can arrive split across TCP segments (seen with
